🩹  (notion) npm build to be from `./dist` rework imports/foldr
