+++ packages/notion/release.config.cjs
@@ -1,6 +1,7 @@
 /* eslint-disable import/order */
 const isCI = require('is-ci')
-!isCI && require('dotenv').config({ path: '../../.env' })
+// @ci(notion) This file is run from `./dist` in build process
+!isCI && require('dotenv').config({ path: '../../../.env' })
 
 const { releaseConfig } = require('@jeromefitz/semantic-config')
 const { name } = require('./package.json')
