+++ packages/notion/src/queries/getInfoType/index.ts
