+++ packages/notion/src/queries/getQuery/index.ts
@@ -2,9 +2,9 @@ import _map from 'lodash/map.js'
 import _omit from 'lodash/omit.js'
 import _size from 'lodash/size.js'
 
-import { PROPERTIES } from '../schema'
-import { avoidRateLimit, isObjectEmpty } from '../utils'
-import dataNormalizedResults from '../utils/dataNormalizedResults'
+import { PROPERTIES } from '../../constants'
+import { avoidRateLimit, isObjectEmpty } from '../../utils'
+import dataNormalizedResults from '../../utils/dataNormalizedResults'
 
 // const useCache = process.env.NEXT_PUBLIC__NOTION_USE_CACHE
 // const useCache = false
