+++ packages/notion/src/utils/isObjectEmpty/index.ts
